+++ kifi-backend/modules/shoebox/app/com/keepit/controllers/core/AuthController.scala
@@ -62,25 +62,6 @@ class AuthController @Inject() (
   // Note: some of the below code is taken from ProviderController in SecureSocial
   // Logout is still handled by SecureSocial directly.
 
-  private implicit val readsOAuth2Info = Json.reads[OAuth2Info]
-
-  def mobileAuth(providerName: String) = Action(parse.json) { implicit request =>
-    // format { "accessToken": "..." }
-    val oauth2Info = request.body.asOpt[OAuth2Info]
-    val provider = Registry.providers.get(providerName).get
-    val filledUser = provider.fillProfile(
-      SocialUser(IdentityId("", provider.id), "", "", "", None, None, provider.authMethod, oAuth2Info = oauth2Info))
-    UserService.find(filledUser.identityId) map { user =>
-      val newSession = Events.fire(new LoginEvent(user)).getOrElse(session)
-      Authenticator.create(user).fold(
-        error => throw error,
-        authenticator => Ok(Json.obj("sessionId" -> authenticator.id))
-          .withSession(newSession - SecureSocial.OriginalUrlKey - IdentityProvider.SessionId - OAuth1Provider.CacheKey)
-          .withCookies(authenticator.toCookie)
-      )
-    } getOrElse NotFound(Json.obj("error" -> "user not found"))
-  }
-
   def loginSocial(provider: String) = ProviderController.authenticate(provider)
   def logInWithUserPass(link: String) = Action { implicit request =>
     ProviderController.authenticate("userpass")(request) match {
